Add parameters to templates
The names and types of the parameters will be included in the
representation of the templates. For example, for a template with an
optional `node_count` parameter the template will be represented like
this (when converted to JSON):

```json
{
  "id": "123",
  "title": "my_template",
  "description": "A simple template that supports setting the number of nodes",
  "parameters": {
    "node_count": {
      "title": "Number of nodes",
      "description": "The number of nodes of the cluster",
      "type": "type.googleapis.com/google.protobuf.Int32Value",
      "default": 3
    }
  }
}
```

The values of the parmaters will be supplied as part of the order. For
example, to create a cluster with 42 nodes:

```json
{
  "order": {
    "spec": {
      "template_id": "123",
      "template_parameters": {
        "node_count": {
	  "@type": "type.googleapis.com/google.protobuf.Int32Value",
	  "value": 42
	}
      }
    }
  }
}
```

Related: https://github.com/innabox/fulfillment-api/issues/9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>